requirements.txt
requests==2.32.3
boto3==1.34.162
snowflake-connector-python==3.12.2
apache-airflow-providers-amazon==8.24.0
apache-airflow-providers-snowflake==5.7.1
